requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
meraki==1.33.0
multidict==6.0.4
numpy==1.24.3
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.30.0
rsa==4.9
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
